requirements/constraints.txt
boto==2.49.0
boto3==1.14.40
botocore==1.17.40
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==3.0
Django==2.2.15
django-appconf==1.0.4
django-axes==5.4.3
django-bootstrap3==14.1.0
django-dbmi-client==0.3.14
django-filter==2.3.0
django-guardian==2.3.0
django-health-check==3.12.1
django-ipware==3.0.1
django-nose==1.4.6
django-taggit==1.3.0
django-taggit-serializer==0.1.7
djangorestframework==3.11.1
djangorestframework-guardian==0.3.0
docutils==0.16
drf-nested-routers==0.91
furl==2.1.0
idna==2.10
importlib-metadata==1.7.0
jmespath==0.10.0
jsonfield==3.1.0
jwcrypto==0.7
mysqlclient==2.0.1
nose==1.3.7
orderedmultidict==1.0.1
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
raven==6.10.0
requests==2.24.0
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
zipp==3.1.0
